Pull out worlds from gpu_voxels_victor into separate files
