+++ src/wip_victor_real.cpp
@@ -1,6 +1,7 @@
 #include <ros/ros.h>
 #include "std_msgs/String.h"
 #include "gpu_voxels_victor.hpp"
+#include "worlds.hpp"
 #include "victor_planning.hpp"
 #include <gpu_voxels/logging/logging_gpu_voxels.h>
 #include <csignal>
